// src/modules/system/user/user.interface.ts
export interface ISystemUserPageQueryRowItem {
  id: number;
  deptId: number;
  jobId: number;
  professionId: number;
  account: string;
  username: string;
  nickname: string;
  avatar: string;
  gender: number;
  professionName: string;
  jobName: string;
  deptName: string;
  roleNames: string;
  roleIds: string;
  email: string;
  mobile: string;
  remark: string;
  orderNum: number;
  status: number;
  createTime: string;
  updateTime: string;
}